+++ src/popup/index.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-import { Popup } from "./component";
+import { Popup } from "./Popup";
 import "../css/app.css";
 
 const mountPopup = (): void => {