+++ src/popup/views/messageEditor/components/colorControl/index.tsx
@@ -1,1 +1,0 @@
-export { ColorControls, ColorControlsProps } from "./component";