+++ src/popup/views/messageEditor/components/colorControls/index.ts
@@ -1,0 +1,5 @@
+export {
+    ColorControls,
+    ColorControlsProps,
+    ColorSetting,
+} from "./ColorControls";