+++ src/popup/views/messageEditor/index.tsx
@@ -1,1 +1,0 @@
-export { MessageEditor, MessageEditorProps } from "./component";